[MOVEONLY] refactor: move CPubKey to pubkey.h
