+++ src/pubkey.cpp
@@ -1,0 +1,6 @@
+// Copyright (c) 2009-2021 The Bitcoin Core developers
+// Copyright (c) 2017 The Zcash developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#include <pubkey.h>
